requirements.txt
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
tensorflow>=2.4.0
xgboost>=1.3.0
joblib>=0.17.0
pyswisseph>=2.0.0
nfl_data_py>=0.3.0
requests>=2.25.1
click>=7.1.2
fastapi>=0.68.0
uvicorn>=0.15.0
python-dateutil>=2.8.1
tqdm>=4.62.0
aiohttp>=3.8.0

# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# Data processing and ML
lightgbm==4.2.0
catboost==1.2.2
imbalanced-learn==0.12.0
scipy==1.11.3
torch==2.1.2
pytorch-lightning==2.1.3

# Visualization
plotly==5.18.0
dash==2.14.2
matplotlib==3.8.2
seaborn==0.13.1

# API and Web
beautifulsoup4==4.12.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Astrological
pyswisseph==2.10.03
ephem==4.1.5
pytz==2023.3.post1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.3
rich==13.7.0

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
